Link tests statically.
https://boringssl-review.googlesource.com/c/12360/ made us define
BORINGSSL_SHARED_LIBRARY when building tests via Bazel. This was to
avoid references to some assembly-defined unexported symbols. This
results in those tests not running and also some nuisances around
Windows because BORINGSSL_SHARED_LIBRARY is not a no-op for consumers on
Windows.

Instead, leave copts alone and link tests statically. (Although this
block is doomed to die anyway since we're close to moving everything to
GTest.)

Change-Id: If237efca219a1f03d64dc801cc1d585556bf2d1d