Update patch set 3
Patch Set 3:

> Patch Set 3:
> 
> > Patch Set 3:
> > 
> > Maybe? Let me first ask a naive question: so far as I can tell this is all about p256-x86_64_test, yes? If the end state is to move that into :crypto_tests, why not just do that now and avoid all this churn?
> 
> Well, if we do that, crypto_tests will then want a linkstatic for the same reason. I figured they'd be equally objectionable. But I can certainly go ahead and move that now. Then we can drop BORINGSSL_SHARED_LIBRARY on create_tests but we'd need to sort out crypto_tests.

It's less objectionable because crypto_tests is not automatically generated, so the knob goes in exactly one place, and I can do stuff like add an explanatory comment, or maybe even make it platform-specific.

> > linkstatic=1 is kind of a hostile thing do, so I'd rather avoid it if/where we can.
> 
> Oh? I'm not very familiar with bazel. What's hostile about it? Note this is just for the test binaries, not any libraries.
> 
> (There's also other options to toy with if linkstatic is objectionable. This just seemed simplest in terms of reducing the number of build variations flying around. If I'm understanding linkstatic right, this matches how the standalone build works.)

Hostile meaning that it is more efficient to use shared libraries.  I also dislike it because it is a knob, and the fewer of those are better.

Admittedly, these are sort of nit-picky, so I don't feel too strongly about them, but I'd still rather have it on crypto_tests if that's possible.

Patch-set: 3
Reviewer: Gerrit User 5645 <5645@66b711a4-aa9b-3dac-9062-916540830212>