Update patch set 11
Patch Set 11:

> Patch Set 11:
> 
> I just realized: this API isn't *quite* enough for TLS 1.3. TLS 1.3 records look like:
> 
>   header ++ ciphertext(plaintext ++ content_type ++ padding) ++ tag
> 
> which means that a consumer hoping to do a TLS-level seal/scatter wants the suffix to be ciphertext(content_type ++ padding) ++ tag.

I think Adam and I briefly talked about this. The discussion was around TLS 1.2, where it's not actually a problem because the padding goes after the tag IIRC. But I think we agreed that "tag" would actually be "everything that's not cipher(plaintext)".

So as long as I clear that aspect up in the docs, we should be good, right?

Patch-set: 11