Don't specialize DeleteImpl on incomplete types.
std::is_trivially_destructable is undefined (C++11) or returns false
(C++17) on incomplete types, neither of which we want. Include sizeof(T)
in the expression to trigger SFINAE and skip template specialization.
Otherwise things get confused.

This requires two layers of enable_if with GCC rather than just &&
because && does not short-circuit template evaluation. The GCC on my
machine implements the C++11 interpretation and fails to compile if
hitting the undefined case.

Bug: 132
Change-Id: I6f2d47499f16e75f62629c76f43a5329e91c6daf