Fix doc.go against Go tip.
Go 1.9 is slated to have some backwards-incompatible changes to
html/template. See https://github.com/golang/go/issues/19952.

If I'm reading this correctly, the issue is that the context-aware auto
escaper would get confused if the text/template compatibility escaper,
html, was used in the wrong context. This does not apply to us because
we never used it in an attribute or so. Nonetheless, we can be
compatible with it and tidy up markupPipeWords' type signature. It
should have had type template.HTML -> template.HTML, not string ->
template.HTML, because it expects the input to be pre-escaped.

I considered adding our own escapeHTML with type string -> template.HTML
and fixing markupPipeWords to be template.HTML -> template.HTML, but
markupPipeWords does not correctly handle all possible template.HTML
input. If a | were in an attribute somewhere, it would mangle the text.
Instead, I kept it of type string -> template.HTML and defined it to
perform the HTML escaping internally. This seems to produce the same
output as before in Go 1.8 and tip.

Change-Id: I90618a3c5525ae54f9fe731352fcff5856b9ba60