Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> > Change has been successfully cherry-picked as c77ea04c81435d6955082d3221857bdfcf71b697 by Adam Langley
> 
> Enabling both avx2 and addx here, while not enabling addx in `x86_64-mont{,5}.pl` means that processors that are capable of both won't use any.
> 
> Tested on E5-2630 v4: `bssl speed` gives ~700 rsa2048, patched with addx=0 in `rsaz-avx2.pl` -- ~1000, patched with addx=1 in `*mont*.pl` -- ~1200.

Oh, thanks! Yeah, it looks like this file only uses $addx to turn *off* rsaz_avx2_eligible, probably because it wants to hit the addx code elsewhere.

Adam, do you have particular opinions as to whether we enable addx elsewhere or disable it here? I'm less scared of enabling random things now that we have the SDE bots, but there were a number of carry bugs in that code recently which we got to ignore. I'm thinking we set $addx to zero here to start with and then evaluate enabling it across the board separately?

Patch-set: 2