Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> > Patch Set 2:
> > 
> > > Patch Set 2:
> > > 
> > > Could we just pass the function pointer instead or will that upset CFI?
> > 
> > It would still be problematic because it's still calling a function pointer cast to a mismatched type signature (since bwrite and bgets/bread have different type signatures.)
> 
> Hrm. Is it also a problem to do things like cast a void (*)(FOO *) to a void (*)(void *) and call it? That's done all over the place.

That's a very common pattern and I am prototyping a change that would 'demote' all FOO*s to void*s to accept this behavior. However, I'm not sure if it's a good idea to remove 'const' qualifiers as would be required for the type mismatch in this case.

Patch-set: 2
Reviewer: Gerrit User 8685 <8685@66b711a4-aa9b-3dac-9062-916540830212>