Update patch set 2
Patch Set 2: Code-Review+2

> Patch Set 1:
> 
> David, if you don't like this change, LMK.

BoringSSL needs to build against 10.12-ish SDK but run on 10.9, so it's all kind of a mess. :-) But our OS requirements are libc which hasn't changed for a while, so this should be fine.

(Plus the standalone CQ is an early warning for the standalone builders, which are just an early warning for the consumer's builders, so minor mismatches aren't critical.)

Patch-set: 2
Reviewer: Gerrit User 5415 <5415@66b711a4-aa9b-3dac-9062-916540830212>
Label: Code-Review=+2