Remove RFC 5114 groups.
These groups are terrible, we got the function wrong (unused ENGINE
parameter does not match upstream), and the functions are unused. Unwind
them. This change doesn't unwind the X9.42 Diffie-Hellman machinery, so
the checks are still present and tested.

(We can probably get rid of them, but it is reachable from DSA_dup_DH.
That's only used by wpa_supplicant and, if that code ever ran, it'd be
ignored because we don't support DHE in TLS. I've left it alone for the
time being.)

Bug: 2
Change-Id: I8d9396983c8d40ed46a03ba6947720da7e9b689a