Update patch set 1
Patch Set 1:

The warning should not be triggering in this case, right? And it's not just that the compiler cannot know that it shouldn't be triggering, it's that the compiler is just being dumb, I think.

If so, dropping the warning for this compiler version in CMake seems like the best answer.

Patch-set: 1
Reviewer: Gerrit User 5005 <5005@66b711a4-aa9b-3dac-9062-916540830212>