Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> You should #include <hmac.h> from src/event/ngx_event_openssl.c if you need it there.

Include What You Use is a pretty well-established rule for C I think, and it'd be nice to gradually trim the amount of stuff ssl.h (and other headers) pull in, so folks don't feel obligated to forward-declare everything all the time. Were the NGINX folks not receptive to fixing this on their end?

Patch-set: 1