DO NOT SUBMIT YET - Store EC_KEY private keys as EC_SCALAR.
[This shouldn't be submitted until
https://boringssl-review.googlesource.com/#/c/boringssl/+/24425 has
stuck, in case it needs to be rolled back. It relies on the
simplification. If it does get rolled back, I can revise this to keep a
pending_priv_key BIGNUM on the EC_KEY, but that's a nuisance, so let's
try the simpler route first.]

This is tricky because EC_KEY_get0_private_key's existence mandates that
an EC_KEY's private key be available as a BIGNUM. So we alias a BIGNUM
into the EC_SCALAR backing store and compute bn->top in constant-time.

Anyone who actually calls EC_KEY_get0_private_key and then passes it to
just about any BIGNUM function will unavoidably leak a side channel, but
the vast majority of consumers who do not do this will not leak.

TODO - check benchmarks real quick

Update-Note: If there are weird bugs around EC_KEY or BIGNUM, this
    change may be at fault. Additionally, some cases where BIGNUMs are
    reused are now created from scratch for simplicity.
Change-Id: I4ce8519c768876ffaf55680ad5c7ef50f1f9a074